requirements.txt
pyqt6 
pyserial 
pandas 
openpyxl